+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.scottyab.rootbeer.sample"
+<manifest package="com.kimchangyoun.rootbeerFresh.sample"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
@@ -7,7 +7,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme.NoActionBar">
         <activity
-            android:name=".MainActivity"
+            android:name="com.kimchangyoun.rootbeerFresh.sample.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             >
