+++ gradle.properties
@@ -18,19 +18,19 @@ org.gradle.jvmargs=-Xmx1536m
 
 VERSION_NAME=0.0.7
 VERSION_CODE=7
-GROUP=com.scottyab
+GROUP=com.kimchangyoun
 
 
 POM_DESCRIPTION=A tasty root checker library and sample app. We've scoured the internets for different methods of answering that age old question... Has this device got root?
-POM_URL=https://github.com/scottyab/rootbeer.git
-POM_SCM_URL=https://github.com/scottyab/rootbeer.git
-POM_SCM_CONNECTION=scm:git@github.com:scottyab/rootbeer.git
-POM_SCM_DEV_CONNECTION=scm:git@github.com:scottyab/rootbeer.git
+POM_URL=https://github.com/kimchangyoun/rootbeerFresh.git
+POM_SCM_URL=https://github.com/kimchangyoun/rootbeerFresh.git
+POM_SCM_CONNECTION=scm:git@github.com:kimchangyoun/rootbeerFresh.git
+POM_SCM_DEV_CONNECTION=scm:git@github.com:kimchangyoun/rootbeerFresh.git
 POM_LICENCE_NAME=The Apache Software License, Version 2.0
 POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
 POM_LICENCE_DIST=repo
-POM_DEVELOPER_ID=scottyab
-POM_DEVELOPER_NAME=Scott Alexander-Bown
+POM_DEVELOPER_ID=kimchangyoun
+POM_DEVELOPER_NAME=KimChangYoun
 
 RELEASE_REPOSITORY_URL=https://oss.sonatype.org/service/local/staging/deploy/maven2
 SNAPSHOT_REPOSITORY_URL=https://oss.sonatype.org/content/repositories/snapshots
