+++ rootbeerlib/gradle.properties
@@ -1,4 +1,4 @@
 # Module Gradle settings.
-POM_NAME=rootbeer
-POM_ARTIFACT_ID=rootbeer-lib
+POM_NAME=rootbeerFresh
+POM_ARTIFACT_ID=rootbeerFresh-lib
 POM_PACKAGING=jar
