+++ rootbeerlib/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.scottyab.rootbeer">
+<manifest package="com.kimchangyoun.rootbeerFresh">
 
     <application />
