+++ rootbeerlib/src/main/java/com/kimchangyoun/rootbeerFresh/Const.java
@@ -1,4 +1,4 @@
-package com.scottyab.rootbeer;
+package com.kimchangyoun.rootbeerFresh;
 
 /**
  * Created by mat on 19/06/15.
