+++ rootbeerlib/src/main/jni/toolChecker.h
@@ -11,10 +11,10 @@ extern "C" {
 
 #include <jni.h>
 
-void Java_com_scottyab_rootbeer_RootBeerNative_setLogDebugMessages( JNIEnv* env, jobject thiz, jboolean debug);
+void Java_com_kimchangyoun_rootbeer_RootBeerNative_setLogDebugMessages( JNIEnv* env, jobject thiz, jboolean debug);
 
-int Java_com_scottyab_rootbeer_RootBeerNative_checkForMagiskUDS( JNIEnv* env, jobject thiz );
+int Java_com_kimchangyoun_rootbeer_RootBeerNative_checkForMagiskUDS( JNIEnv* env, jobject thiz );
 
-int Java_com_scottyab_rootbeer_RootBeerNative_checkForRoot( JNIEnv* env, jobject thiz , jobjectArray pathsArray );
+int Java_com_kimchangyoun_rootbeer_RootBeerNative_checkForRoot( JNIEnv* env, jobject thiz , jobjectArray pathsArray );
 
 }
