+++ rootbeerlib/src/test/java/com/kimchangyoun/rootbeerFresh/RootBeerTest.java
@@ -1,4 +1,4 @@
-package com.scottyab.rootbeer;
+package com.kimchangyoun.rootbeerFresh;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
